+++ src/inference/dev_api/openvino/runtime/threading/cpu_streams_executor.hpp
@@ -3,37 +3,35 @@
 //
 
 /**
- * @file ie_cpu_streams_executor.hpp
- * @brief A header file for Inference Engine CPU-Streams-based Executor implementation.
+ * @file openvino/runtime/threading/cpu_streams_executor.hpp
+ * @brief A header file for OpenVINO CPU-Streams-based Executor implementation.
  */
 
 #pragma once
 
 #include <memory>
+#include <openvino/runtime/common.hpp>
 #include <string>
 
-#include "threading/ie_istreams_executor.hpp"
+#include "openvino/runtime/threading/istreams_executor.hpp"
 
-namespace InferenceEngine {
+namespace ov {
+namespace threading {
+
 /**
  * @class CPUStreamsExecutor
- * @ingroup ie_dev_api_threading
+ * @ingroup ov_dev_api_threading
  * @brief CPU Streams executor implementation. The executor splits the CPU into groups of threads,
  *        that can be pinned to cores or NUMA nodes.
  *        It uses custom threads to pull tasks from single queue.
  */
-class INFERENCE_ENGINE_API_CLASS(CPUStreamsExecutor) : public IStreamsExecutor {
+class OPENVINO_RUNTIME_API CPUStreamsExecutor : public IStreamsExecutor {
 public:
-    /**
-     * @brief A shared pointer to a CPUStreamsExecutor object
-     */
-    using Ptr = std::shared_ptr<CPUStreamsExecutor>;
-
     /**
      * @brief Constructor
      * @param config Stream executor parameters
      */
-    explicit CPUStreamsExecutor(const InferenceEngine::IStreamsExecutor::Config& config = {});
+    explicit CPUStreamsExecutor(const Config& config);
 
     /**
      * @brief A class destructor
@@ -42,15 +40,16 @@ public:
 
     void run(Task task) override;
 
-    void Execute(Task task) override;
+    void execute(Task task) override;
 
-    int GetStreamId() override;
+    int get_stream_id() override;
 
-    int GetNumaNodeId() override;
+    int get_numa_node_id() override;
 
 private:
     struct Impl;
     std::unique_ptr<Impl> _impl;
 };
 
-}  // namespace InferenceEngine
+}  // namespace threading
+}  // namespace ov
