+++ src/inference/src/dev/iasync_infer_request.cpp
@@ -14,13 +14,13 @@
 
 namespace {
 
-struct ImmediateStreamsExecutor : public InferenceEngine::ITaskExecutor {
-    explicit ImmediateStreamsExecutor(const InferenceEngine::IStreamsExecutor::Ptr& streamsExecutor)
+struct ImmediateStreamsExecutor : public ov::threading::ITaskExecutor {
+    explicit ImmediateStreamsExecutor(const std::shared_ptr<ov::threading::IStreamsExecutor>& streamsExecutor)
         : _streamsExecutor{streamsExecutor} {}
     void run(InferenceEngine::Task task) override {
-        _streamsExecutor->Execute(std::move(task));
+        _streamsExecutor->execute(std::move(task));
     }
-    InferenceEngine::IStreamsExecutor::Ptr _streamsExecutor;
+    std::shared_ptr<ov::threading::IStreamsExecutor> _streamsExecutor;
 };
 
 }  // namespace
@@ -30,8 +30,8 @@ ov::IAsyncInferRequest::~IAsyncInferRequest() {
 }
 
 ov::IAsyncInferRequest::IAsyncInferRequest(const std::shared_ptr<IInferRequest>& request,
-                                           const InferenceEngine::ITaskExecutor::Ptr& task_executor,
-                                           const InferenceEngine::ITaskExecutor::Ptr& callback_executor)
+                                           const std::shared_ptr<ov::threading::ITaskExecutor>& task_executor,
+                                           const std::shared_ptr<ov::threading::ITaskExecutor>& callback_executor)
     : m_sync_request(request),
       m_request_executor(task_executor),
       m_callback_executor(callback_executor) {
@@ -117,7 +117,7 @@ void ov::IAsyncInferRequest::start_async_thread_unsafe() {
 
 void ov::IAsyncInferRequest::run_first_stage(const Pipeline::iterator itBeginStage,
                                              const Pipeline::iterator itEndStage,
-                                             const InferenceEngine::ITaskExecutor::Ptr callbackExecutor) {
+                                             const std::shared_ptr<ov::threading::ITaskExecutor> callbackExecutor) {
     auto& firstStageExecutor = std::get<Stage_e::EXECUTOR>(*itBeginStage);
     OPENVINO_ASSERT(nullptr != firstStageExecutor);
     firstStageExecutor->run(make_next_stage_task(itBeginStage, itEndStage, std::move(callbackExecutor)));
@@ -126,9 +126,9 @@ void ov::IAsyncInferRequest::run_first_stage(const Pipeline::iterator itBeginSta
 InferenceEngine::Task ov::IAsyncInferRequest::make_next_stage_task(
     const Pipeline::iterator itStage,
     const Pipeline::iterator itEndStage,
-    const InferenceEngine::ITaskExecutor::Ptr callbackExecutor) {
+    const std::shared_ptr<ov::threading::ITaskExecutor> callbackExecutor) {
     return std::bind(
-        [this, itStage, itEndStage](InferenceEngine::ITaskExecutor::Ptr& callbackExecutor) mutable {
+        [this, itStage, itEndStage](std::shared_ptr<ov::threading::ITaskExecutor>& callbackExecutor) mutable {
             std::exception_ptr currentException = nullptr;
             auto& thisStage = *itStage;
             auto itNextStage = itStage + 1;
