+++ src/inference/src/dev/iplugin.cpp
@@ -4,7 +4,7 @@
 
 #include "openvino/runtime/iplugin.hpp"
 
-ov::IPlugin::IPlugin() : m_executor_manager(ov::executor_manager()), m_is_new_api(true) {}
+ov::IPlugin::IPlugin() : m_executor_manager(ov::threading::executor_manager()), m_is_new_api(true) {}
 
 void ov::IPlugin::set_version(const ov::Version& version) {
     m_version = version;
@@ -42,7 +42,7 @@ bool ov::IPlugin::is_new_api() const {
     return m_is_new_api;
 }
 
-const std::shared_ptr<ov::ExecutorManager>& ov::IPlugin::get_executor_manager() const {
+const std::shared_ptr<ov::threading::ExecutorManager>& ov::IPlugin::get_executor_manager() const {
     return m_executor_manager;
 }
