+++ src/inference/tests/unit/ie_executor_manager_tests.cpp
@@ -4,36 +4,34 @@
 
 #include <gtest/gtest.h>
 
-#include <threading/ie_executor_manager.hpp>
+#include "openvino/runtime/threading/executor_manager.hpp"
 
 using namespace ::testing;
-using namespace std;
-using namespace InferenceEngine;
 
 TEST(ExecutorManagerTests, canCreateSingleExecutorManager) {
-    auto executorManager1 = executorManager();
+    auto executorManager1 = ov::threading::executor_manager();
 
-    auto executorManager2 = executorManager();
+    auto executorManager2 = ov::threading::executor_manager();
 
     ASSERT_EQ(executorManager1, executorManager2);
 }
 
 TEST(ExecutorManagerTests, createDifferentExecutorsForDifferentDevices) {
-    auto executorMgr = executorManager();
-    auto executor1 = executorMgr->getExecutor("CPU");
-    auto executor2 = executorMgr->getExecutor("GPU");
+    auto executorMgr = ov::threading::executor_manager();
+    auto executor1 = executorMgr->get_executor("CPU");
+    auto executor2 = executorMgr->get_executor("GPU");
 
     ASSERT_NE(executor1, executor2);
-    ASSERT_EQ(2, executorMgr->getExecutorsNumber());
+    ASSERT_EQ(2, executorMgr->get_executors_number());
 }
 
 TEST(ExecutorManagerTests, returnTheSameExecutorForTheSameDevice) {
-    auto executorMgr = executorManager();
-    auto executor1 = executorMgr->getExecutor("CPU");
-    auto executor2 = executorMgr->getExecutor("GPU");
+    auto executorMgr = ov::threading::executor_manager();
+    auto executor1 = executorMgr->get_executor("CPU");
+    auto executor2 = executorMgr->get_executor("GPU");
 
-    auto executor = executorMgr->getExecutor("GPU");
+    auto executor = executorMgr->get_executor("GPU");
 
     ASSERT_EQ(executor, executor2);
-    ASSERT_EQ(2, executorMgr->getExecutorsNumber());
+    ASSERT_EQ(2, executorMgr->get_executors_number());
 }
