+++ src/plugins/template/src/template_config.cpp
@@ -16,16 +16,17 @@ Configuration::Configuration() {}
 Configuration::Configuration(const ConfigMap& config, const Configuration& defaultCfg, bool throwOnUnsupported) {
     *this = defaultCfg;
     // If plugin needs to use InferenceEngine::StreamsExecutor it should be able to process its configuration
-    auto streamExecutorConfigKeys = _streamsExecutorConfig.SupportedKeys();
+    auto streamExecutorConfigKeys =
+        _streamsExecutorConfig.get_property(ov::supported_properties.name()).as<std::vector<std::string>>();
     for (auto&& c : config) {
         const auto& key = c.first;
         const auto& value = c.second;
 
         if (ov::template_plugin::throughput_streams == key) {
-            _streamsExecutorConfig.SetConfig(CONFIG_KEY(CPU_THROUGHPUT_STREAMS), value.as<std::string>());
+            _streamsExecutorConfig.set_property(CONFIG_KEY(CPU_THROUGHPUT_STREAMS), value);
         } else if (streamExecutorConfigKeys.end() !=
                    std::find(std::begin(streamExecutorConfigKeys), std::end(streamExecutorConfigKeys), key)) {
-            _streamsExecutorConfig.SetConfig(key, value.as<std::string>());
+            _streamsExecutorConfig.set_property(key, value);
         } else if (CONFIG_KEY(DEVICE_ID) == key) {
             deviceId = std::stoi(value.as<std::string>());
             if (deviceId > 0) {
@@ -42,11 +43,12 @@ Configuration::Configuration(const ConfigMap& config, const Configuration& defau
     }
 }
 
-InferenceEngine::Parameter Configuration::Get(const std::string& name) const {
-    auto streamExecutorConfigKeys = _streamsExecutorConfig.SupportedKeys();
+ov::Any Configuration::Get(const std::string& name) const {
+    auto streamExecutorConfigKeys =
+        _streamsExecutorConfig.get_property(ov::supported_properties.name()).as<std::vector<std::string>>();
     if ((streamExecutorConfigKeys.end() !=
          std::find(std::begin(streamExecutorConfigKeys), std::end(streamExecutorConfigKeys), name))) {
-        return _streamsExecutorConfig.GetConfig(name);
+        return _streamsExecutorConfig.get_property(name);
     } else if (name == CONFIG_KEY(DEVICE_ID)) {
         return {std::to_string(deviceId)};
     } else if (name == CONFIG_KEY(PERF_COUNT)) {
@@ -54,7 +56,7 @@ InferenceEngine::Parameter Configuration::Get(const std::string& name) const {
     } else if (name == ov::template_plugin::throughput_streams || name == CONFIG_KEY(CPU_THROUGHPUT_STREAMS)) {
         return {std::to_string(_streamsExecutorConfig._streams)};
     } else if (name == CONFIG_KEY(CPU_BIND_THREAD)) {
-        return const_cast<InferenceEngine::IStreamsExecutor::Config&>(_streamsExecutorConfig).GetConfig(name);
+        return _streamsExecutorConfig.get_property(name);
     } else if (name == CONFIG_KEY(CPU_THREADS_NUM)) {
         return {std::to_string(_streamsExecutorConfig._threads)};
     } else if (name == CONFIG_KEY_INTERNAL(CPU_THREADS_PER_STREAM)) {
