+++ src/plugins/template/src/template_config.hpp
@@ -4,11 +4,11 @@
 
 #pragma once
 
-#include <ie_parameter.hpp>
 #include <map>
-#include <openvino/runtime/properties.hpp>
 #include <string>
-#include <threading/ie_istreams_executor.hpp>
+
+#include "openvino/runtime/properties.hpp"
+#include "openvino/runtime/threading/istreams_executor.hpp"
 
 namespace TemplatePlugin {
 
@@ -26,13 +26,13 @@ struct Configuration {
                            const Configuration& defaultCfg = {},
                            const bool throwOnUnsupported = true);
 
-    InferenceEngine::Parameter Get(const std::string& name) const;
+    ov::Any Get(const std::string& name) const;
 
     // Plugin configuration parameters
 
     int deviceId = 0;
     bool perfCount = true;
-    InferenceEngine::IStreamsExecutor::Config _streamsExecutorConfig;
+    ov::threading::IStreamsExecutor::Config _streamsExecutorConfig;
     ov::hint::PerformanceMode performance_mode = ov::hint::PerformanceMode::UNDEFINED;
 };
 // ! [configuration:header]
